[SDK-686] idtype and request t (#39)
* tests, id type and request t added

* changelog entry

* linting

* more linting

* some timing issue fix

* update

* wording changes

* changelog and split up the init id assign functions

* some comments

* ID type ifs

* tests, comments and linting

* timing for tests

+++ test/helpers/helper_functions.js
@@ -25,18 +25,21 @@ function readRequestQueue() {
 }
 
 // queue files clearing logic
-function clearStorage() {
+function clearStorage(keepID) {
+    keepID = keepID || false;
     // Resets Countly
     Countly.halt(true);
     // clean storages
-    if (fs.existsSync(idDir)) {
-        fs.unlinkSync(idDir);
-    }
     if (fs.existsSync(eventDir)) {
         fs.unlinkSync(eventDir);
     }
     if (fs.existsSync(reqDir)) {
         fs.unlinkSync(reqDir);
+    }
+    if (!keepID) {
+        if (fs.existsSync(idDir)) {
+            fs.unlinkSync(idDir);
+        }
     }
 }
 /**
